split ai cog into image_generation and text_generation
